requirements.txt
flask>=3.0
python-dotenv>=1.0
groq>=0.9
python-docx>=1.1
duckduckgo-search>=7.0
reportlab>=4.1
lxml>=5.0
gunicorn>=21.0
